Support relocatable doc sets
The symbol now store a path relative to the doc set URL and the doc set
URL is a file reference URL. This allows the doc set to be moved/renamed
while the app is running or between app launches and the path to search
results will continue to be resolved without a need for reindexing.

Not dealing with NSURLs for each symbol also significantly speeds up
loading of the cached index.