Add fast but evil database indexer
This indexer is evil because it reads directly from the SQLite Core Data
database that serves as the doc set's index. However, in some cases this
is the only good source of API information. For example, appledoc does
not include anchors for any of its classes or protocols, resulting in
the HTML indexer extracting very little useful information from such doc
sets.

The database indexer is also much faster than HTML indexing, to the
point of being able to extract all API tokens from a platform doc set in
under a second.